Refactor unit tests and helper methods
- Refactors the helper functions for bidirectionalizing and reversing
graphs, so that they do not mutate the input graphs but rather return
new graphs. (Mutating function arguments is a bad smell in general, and
should be avoided when not necessary.)

- Create a new helper base test class with assertions for graph
isomorphism and that a graph is isomorphic to a provided iterator of
edges.

- Write more extensive tests for the bidirectional and reverse
functions.

Test plan: `python3 infra/graph_generators_test.py`