Add Jupyter nbextension library for research branch data visualization (#27)
* first graph version - pre-EDA

* graphNet updates for further abstraction

* initialize simulation; abstract label; add documentation

* convert script to library with scaffolding

* add source cred eda on graph data

* edit annoying spelling/copy&paste errors in eda

* add d3-graphNet to Jupyter Notebook

* scaffold d3 dep and python-JS interface

* get d3 chart into Jupyter example

* add attachGraphNet function; remove garbage notebook

* remove notebooks; add python-JS widget

widget still throws an error that I can't debug. Will make PR when I debug

* working widget added; garbage out

Finally made it through all the dumpster fires for this one. Still need to make working Python functions for sub graph inspectors.

* basic library ready for pr

Ok, the usage is pretty basic and the chart a bit rough, but it should be a good foundation for working in Jupyter notebooks

* move to infra for PR

putting the library into the right place

* alter JS to allow better Python code usage

Prior to this commit, a force directed chart would have to be initialized in one block without drawing anything, then updated in the next block. This isn't good for writing code. With this commit, a user can initialize the chart completely in a single code block. Demo in `basic_usage` notebook updated as well.

* Update basic_usage.ipynb

* Revert "Update basic_usage.ipynb"

This reverts commit 4933c17a128ddac4d2a9d99008095da6cf2c0ab4.

* Revert "alter JS to allow better Python code usage"

This reverts commit c0a2d86d774942e73b13cff0d99e4d4af27bb92e.

* Revert "move to infra for PR"

This reverts commit 5ae718b7260caaf8fcb9c66da23ff7f77cae5379.

* Revert "basic library ready for pr"

This reverts commit f1b8ef39063f82313127b472e0096be6e5b28558.

* Python code for 1/3 PRs

* Update setup.py